+++ Assets/Scripts/Enemy/IEnemy.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public interface IEnemy {
     public float MovementSpeed { get; set; }
-    public int Health { get; set; }
-    public int MaxHealth { get; set; }
+    public float Health { get; set; }
+    public float MaxHealth { get; set; }
     public int Value { get; set; }
 
     public bool Alive { get; set; }
